requirements.txt
numpy
opencv-python
pyautogui
keyboard
Pillow
ultralytics
pywin32
